backend-api/requirements.txt
alembic==1.18.3
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.2
cffi==2.0.0
click==8.2.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.3.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.47.2
tomli==2.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
